+++ R/CFAxisDiscrete.R
@@ -9,21 +9,24 @@
 CFAxisDiscrete <- R6::R6Class("CFAxisDiscrete",
   inherit = CFAxis,
   private = list(
+    # The values of the axis: just an integer sequence
     get_values = function() {
       seq(self$length)
     },
 
+    # The coordinate values of the axis are just an integer sequence, unless
+    # labels have been set for the axis.
+    get_coordinates = function() {
+      crds <- self$label_set() # Get labels, if they are active
+      if (is.null(crds)) seq(self$length)
+      else crds
+    },
+
     dimvalues_short = function() {
-      lbls <- self$labels
-      if (!length(lbls)) {
-        values <- if (self$length == 1L) "[1]"
-        else paste0("[1 ... ", self$length, "]")
-      } else {
-        lbls <- lbls[[1L]]$values
-        values <- if (self$length == 1L) paste0("[", lbls[1L], "]")
-        else paste0("[", lbls[1L], " ... ", lbls[length(lbls)], "]")
-      }
-      values
+      crds <- private$get_coordinates()
+      len <- self$length
+      if (len == 1L) paste0("[", crds[1L], "]")
+      else paste0("[", crds[1L], " ... ", crds[len], "]")
     }
   ),
   public = list(
@@ -61,6 +64,47 @@ CFAxisDiscrete <- R6::R6Class("CFAxisDiscrete",
       x[x < 1] <- 0L
       x[x > self$length] <- .Machine$integer.max
       as.integer(x)
+    },
+
+    #' @description Return an axis spanning a smaller coordinate range. This
+    #'   method returns an axis which spans the range of indices given by the
+    #'   `rng` argument.
+    #'
+    #' @param group The group to create the new axis in.
+    #' @param rng The range of indices from this axis to include in the returned
+    #'   axis.
+    #'
+    #' @return A `CFAxisDiscrete` instance covering the indicated range of
+    #'   indices. If the `rng` argument includes only a single value, an
+    #'   [CFAxisScalar] instance is returned with the value from this axis. If
+    #'   the value of the argument is `NULL`, return the entire axis (possibly
+    #'   as a scalar axis).
+    sub_axis = function(group, rng = NULL) {
+      var <- NCVariable$new(-1L, self$name, group, "NC_DOUBLE", 1L, NULL)
+
+      .make_scalar <- function(idx) {
+        scl <- CFAxisScalar$new(group, var, self$orientation, idx)
+        private$copy_label_subset_to(scl, idx)
+        scl
+      }
+
+      if (is.null(rng)) {
+        if (self$length > 1L) {
+          ax <- self$clone()
+          ax$group <- group
+          ax
+        } else
+          .make_scalar(1L)
+      } else {
+        if (rng[1L] == rng[2L])
+          .make_scalar(rng[1L])
+        else {
+          dim <- NCDimension$new(-1L, self$name, rng[2L] - rng[1L] + 1L, FALSE)
+          discr <- CFAxisDiscrete$new(group, var, dim, self$orientation)
+          private$copy_label_subset_to(scl, idx)
+          discr
+        }
+      }
     }
   ),
   active = list(
@@ -73,10 +117,8 @@ CFAxisDiscrete <- R6::R6Class("CFAxisDiscrete",
     #' @field dimnames (read-only) The coordinates of the axis as an integer
     #' vector, or labels for every axis element if they have been set.
     dimnames = function(value) {
-      if (missing(value)) {
-        if (length(self$lbls)) self$lbls[[1L]]$values
-        else seq(self$length)
-      }
+      if (missing(value))
+        private$get_coordinates()
     }
   )
 )