+++ R/CFAxisLatitude.R
@@ -39,6 +39,7 @@ CFAxisLatitude <- R6::R6Class("CFAxisLatitude",
         scl <- CFAxisScalar$new(group, var, "Y", idx)
         bnds <- self$bounds
         if (inherits(bnds, "CFBounds")) scl$bounds <- bnds$sub_bounds(group, idx)
+        private$copy_label_subset_to(scl, idx)
         scl
       }
 
@@ -57,6 +58,7 @@ CFAxisLatitude <- R6::R6Class("CFAxisLatitude",
           lat <- CFAxisLatitude$new(group, var, dim, self$values[rng[1L]:rng[2L]])
           bnds <- self$bounds
           if (inherits(bnds, "CFBounds")) lat$bounds <- bnds$sub_bounds(group, rng)
+          private$copy_label_subset_to(lat, idx)
           lat
         }
       }